OverridingConstructor-AddingAttributes-Superfunction
Here is the code we can see the Overriding construction which is to change the purpose of something. There is a Parent class and a child class here. I used the overview to add attributes of the child class. It can be seen in the part of the child class that I used "super().__init__(attributes). Super() can only be used by a child class and it pertains to its parent using that function we can access every function in our parent class inside our child class. So it's like we're just calling the child class. In our code, we can see that super() is calling the parent class and running the .__init__ so that we can still retain its constructor. but we can still add attributes under it.